created separate tests and used in TestXdsLoadBalancerWithFacilitiesFactory and XdsClientOtelMetricsProviderTest.java
